# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MVP multi-tenant event platform. Auth owns user identity; org membership and permissions live in organization/iam.",
        "title": "Rallya API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/orgs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Newest first with member counts. Audited per call.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Platform org inventory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name/slug search",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "Page (1-based)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 20,
                        "description": "Page size, max 100",
                        "name": "perPage",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admindto.OrgsPage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/admindto.ErrorAlias"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/admindto.ErrorAlias"
                        }
                    }
                }
            }
        },
        "/admin/orgs/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Org row with full roster (identity + roles).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Tenant dispute-triage view",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Org UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admindto.OrgDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/admindto.ErrorAlias"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/admindto.ErrorAlias"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/admindto.ErrorAlias"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/admindto.ErrorAlias"
                        }
                    }
                }
            }
        },
        "/admin/orgs/{id}/policies/reseed": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Idempotent re-seed; returns what changed. Audited.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Repair org Casbin rows",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Org UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admindto.PolicyDiff"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/admindto.ErrorAlias"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/admindto.ErrorAlias"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/admindto.ErrorAlias"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/admindto.ErrorAlias"
                        }
                    }
                }
            }
        },
        "/admin/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Email-fragment search with superadmin flags.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Platform account lookup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email fragment",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "Page (1-based)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 20,
                        "description": "Page size, max 100",
                        "name": "perPage",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admindto.UsersPage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/admindto.ErrorAlias"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/admindto.ErrorAlias"
                        }
                    }
                }
            }
        },
        "/admin/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Profile plus every tenant membership.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Account access-review view",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admindto.UserDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/admindto.ErrorAlias"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/admindto.ErrorAlias"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/admindto.ErrorAlias"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/admindto.ErrorAlias"
                        }
                    }
                }
            }
        },
        "/admin/users/{id}/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "orderdto shapes (already Stripe-free). Fraud/refund investigation.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Cross-tenant order history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "Page (1-based)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 20,
                        "description": "Page size, max 100",
                        "name": "perPage",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/orderdto.OrdersPage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/admindto.ErrorAlias"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/admindto.ErrorAlias"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/admindto.ErrorAlias"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/admindto.ErrorAlias"
                        }
                    }
                }
            }
        },
        "/admin/users/{id}/policies/sync": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sweep-then-add per org; stale orgs swept; g2 untouched. Audited.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Converge user groupings to memberships",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admindto.PolicyDiff"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/admindto.ErrorAlias"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/admindto.ErrorAlias"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/admindto.ErrorAlias"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/admindto.ErrorAlias"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "admindto.ErrorAlias": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "organization not found"
                }
            }
        },
        "admindto.MemberRow": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "jane@test.com"
                },
                "emailVerified": {
                    "type": "boolean",
                    "example": true
                },
                "joinedAt": {
                    "type": "string",
                    "example": "2026-09-20T10:00:00Z"
                },
                "name": {
                    "type": "string",
                    "example": "Jane"
                },
                "role": {
                    "type": "string",
                    "example": "ADMIN"
                },
                "userId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "admindto.MembershipRow": {
            "type": "object",
            "properties": {
                "joinedAt": {
                    "type": "string",
                    "example": "2026-09-20T10:00:00Z"
                },
                "orgId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "orgName": {
                    "type": "string",
                    "example": "Acme"
                },
                "orgSlug": {
                    "type": "string",
                    "example": "acme"
                },
                "role": {
                    "type": "string",
                    "example": "ADMIN"
                }
            }
        },
        "admindto.OrgDetail": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2026-09-20T10:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/admindto.MemberRow"
                    }
                },
                "name": {
                    "type": "string",
                    "example": "Acme"
                },
                "slug": {
                    "type": "string",
                    "example": "acme"
                }
            }
        },
        "admindto.OrgSummary": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2026-09-20T10:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "members": {
                    "type": "integer",
                    "example": 12
                },
                "name": {
                    "type": "string",
                    "example": "Acme"
                },
                "slug": {
                    "type": "string",
                    "example": "acme"
                }
            }
        },
        "admindto.OrgsPage": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/admindto.OrgSummary"
                    }
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "perPage": {
                    "type": "integer",
                    "example": 20
                },
                "total": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "admindto.PolicyDiff": {
            "type": "object",
            "properties": {
                "added": {
                    "type": "integer",
                    "example": 2
                },
                "removed": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 27
                }
            }
        },
        "admindto.UserDetail": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2026-09-20T10:00:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "jane@test.com"
                },
                "emailVerified": {
                    "type": "boolean",
                    "example": true
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "memberships": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/admindto.MembershipRow"
                    }
                },
                "name": {
                    "type": "string",
                    "example": "Jane"
                },
                "superAdmin": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "admindto.UserSummary": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2026-09-20T10:00:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "jane@test.com"
                },
                "emailVerified": {
                    "type": "boolean",
                    "example": true
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "name": {
                    "type": "string",
                    "example": "Jane"
                },
                "superAdmin": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "admindto.UsersPage": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/admindto.UserSummary"
                    }
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "perPage": {
                    "type": "integer",
                    "example": 20
                },
                "total": {
                    "type": "integer",
                    "example": 7
                }
            }
        },
...
        "orderdto.OrdersPage": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/orderdto.Order"
                    }
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "perPage": {
                    "type": "integer",
                    "example": 20
                },
                "total": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer \u003caccessToken\u003e from POST /auth/login (or /auth/refresh).\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}